Initial models.py
- Created Django intern admin interface
- Created the base superuser
- Created the Jogador Class:
( This is an secondary user class. that will work as the login model of the players. It uses the login, register and painel URL's)
- Created the /login URL:
(Render an login interface for the Jogador`s users)
- Created the /register URL:
(Render an form to create an Jogador user)
- Created an /pianel URL:
(Renders the basic info of the player. This is just an placeholder to be used as future reference to get player data and eventually the player game screen)

- Added the migrate to the makefile:
(This is required to be runned after each model changes as this is what makes the modifications on the database)

- Added the run-django command to the makefile:
(This is just a shortcut to start the server)